move review and internal status to shared
