--- backend/functions/src/models/appStatus.ts
@@ -1,20 +1,0 @@
-import type { ApplicantRole } from "@app-portal/shared/constants";
-
-export enum ReviewStatus {
-  NotReviewed = "not-reviewed",
-  UnderReview = "under-review",
-  Reviewed = "reviewed",
-  Interview = "interview",
-  Accepted = "accepted",
-  Denied = "denied",
-  Waitlisted = "waitlist",
-}
-
-export type InternalApplicationStatus = {
-  id: string;
-  formId: string;
-  role: ApplicantRole;
-  responseId: string;
-  status: ReviewStatus;
-  isQualified: boolean;
-};
